refactor(web): split the five near-limit size-ratchet functions
Removes five max-lines-per-function suppressions by moving one section
out of each function instead of widening the limit:

- Security and the admin TOTP panel rendered the same recovery-codes
  reveal; both now use a shared TotpRecoveryCodes component.
- useArenaRunner's streaming body is a plain async function in
  streamModel.ts that takes the mount-gated setters as a context object.
- The dashboard's global and per-section range/metric selections, with
  the header-to-section sync effects, live in useDashboardRanges.
- The apprise reachability indicator on the alerts card is its own
  AppriseStatus component; the reason-code set moved next to the API
  text helpers it belongs with.
- The six dashboard stat tiles are a StatCardsRow component typed from
  the useDashboard return.

No behaviour change; existing page tests render the moved JSX.
